# ood-gen: call crunch on each directory separately