This has several benefits:
- the various directories can be processed in parallel
- the generated modules are smaller, so the superlinear behavior of the
  compiler is less visible
- dependencies are more precise, so fewer recompilations should be
  necessary