@emillon 2024-01-22
This PR is more to start a discussion. While contributing I noticed that a single command was taking a long time. The ocaml compiler has trouble with large generated modules. So instead of doing that, this PR generates smaller modules. They build more efficiently, can be built in parallel and have better incremental properties.

Here's a quick benchmark:
- this PR: dune build  51.69s user 7.19s system 318% cpu 18.475 total
- main: dune build  71.25s user 7.92s system 200% cpu 39.436 total
(so we can see that there's both less computation to do, and it is also more parallel)

There are several places we can go with that:
- we can make a dune rule generator to remove some boilerplate and make the code easier to adapt when the directory hierarchy will change
- we could experiment with different scheduling strategies to make the data buckets more even (for example, add more layers of dispatch, or dispatch based on the filename hash)
- maybe this should be upstreamed to crunch or a new tool

Let me know what you think.